requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
Django==1.11.12
django-rest-framework==0.1.0
djangorestframework==3.8.2
feedparser==5.2.1
kombu==4.1.0
pytz==2018.4
vine==1.1.4
